feat: implement OAuth usecase pattern
